refactor(toolbar): Move rename/save props to sidebar
The `Toolbar` component was receiving props related to file renaming and saving, which are more relevant to the file interaction within the sidebar. These props have been moved to the `Sidebar` component to better align with their functional context. The `Toolbar` now receives an `activeView` prop to conditionally render certain elements.